Developing Force DimenFix

+++ dimenfix/log.py
@@ -3,26 +3,28 @@ import time
 import os
 
 
-def clean():
-    os.makedirs("./log", exist_ok=True)
+def clean(active=True):
+    if active is True:
+        os.makedirs("./log", exist_ok=True)
 
-    for filename in os.listdir("./log"):
-        if filename.endswith('.png'):
-            os.remove(os.path.join("./log", filename))
+        for filename in os.listdir("./log"):
+            if filename.endswith('.png'):
+                os.remove(os.path.join("./log", filename))
 
 
-def print_layout(y, label, title=None):
-    plt.figure()
-    plt.scatter(y[:, 0], y[:, 1], c=label,
-                cmap='tab10', edgecolors='face', linewidths=0.5, s=12)
-    plt.grid(linestyle='dotted')
+def print_layout(y, label, title=None, active=True):
+    if active is True:
+        plt.figure()
+        plt.scatter(y[:, 0], y[:, 1], c=label,
+                    cmap='tab10', edgecolors='face', linewidths=0.5, s=12)
+        plt.grid(linestyle='dotted')
 
-    if title is not None:
-        plt.title(title)
+        if title is not None:
+            plt.title(title)
 
-    os.makedirs("./log", exist_ok=True)
+        os.makedirs("./log", exist_ok=True)
 
-    timestamp = int(time.time() * 1000.0)
-    filename = "./log/log_print_layout_" + str(timestamp) + ".png"
-    plt.savefig(filename, dpi=150, bbox_inches='tight')
-    plt.close()
+        timestamp = int(time.time() * 1000.0)
+        filename = "./log/log_print_layout_" + str(timestamp) + ".png"
+        plt.savefig(filename, dpi=150, bbox_inches='tight')
+        plt.close()
